+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - oneview_logical_enclosure
 - oneview_logical_interconnect_group
 - oneview_network_set
+- oneview-scope
 - oneview_server_hardware
 - oneview_server_hardware_type
 - oneview_server_profile